runtime/vm: extract query compilation into vm_compile_query.go
Pure code motion. vm_compile_query.go now carries compileQuery,
compileQueryFull, compileJoins, the hash/single-row/empty join family,
compileGroupQuery and its any/accum/from/join variants, buildRowMap,
and compileQueryFrom. vm.go retains the rest of the compiler.

vm.go: 4147 lines (down from 7042).
vm_compile_query.go: 2904 lines.

Stacked on #21381.

Refs #21382
